v16.5.0 build 42 NOTES - Developing: Database support - Fixed: some issues with previous build relating to trinkets
- NOTE: Still needs some yml attribute updating

(not intended for usage, only testing)

+++ randompackage-spigot/src/main/java/me/randomhashtags/randompackage/api/Trinkets.java
@@ -98,7 +98,7 @@ public class Trinkets extends EventAttributes implements RPItemStack {
             }
         }
     }
-    private int didTriggerTrinket(Event event, ItemStack is, Player player) {
+    private byte didTriggerTrinket(Event event, ItemStack is, Player player) {
         final String id = getRPItemStackValue(is, "TrinketInfo");
         if(id != null) {
             final String[] info = id.split(":");
